pyproject.toml
[tool.poetry]
name = "scenorita-v3"
version = "0.1.0"
description = ""
authors = ["Y <yhuai@uci.edu>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
loguru = "^0.7.0"
absl-py = "^1.4.0"
cyber-record = "^0.1.12"
numpy = "^1.24.2"
pandas = "^2.0.0"
shapely = "^2.0.1"
matplotlib = "^3.7.1"
networkx = "^3.1"
docker = "^6.0.1"
nanoid = "^2.0.0"
deap = "^1.3.3"


[tool.poetry.group.dev.dependencies]
mypy = "^1.2.0"
black = "^23.3.0"
isort = "^5.12.0"
ruff = "^0.0.261"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
ignore_missing_imports = true
exclude = ['modules', 'tests']

[tool.isort]
profile = "black"

[tool.ruff]
exclude = ['modules', 'tests']
